testing/web-platform/tests/websockets/interfaces/WebSocket/close/close-connecting.html

<!doctype html>
<title>WebSockets: close() when connecting</title>
<meta name=timeout content=long>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js?pipe=sub></script>
<div id=log></div>
<script>
var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/sleep_10_v13');
setTimeout(function() {
  assert_equals(ws.readyState, ws.CONNECTING);
  ws.close();
  assert_equals(ws.readyState, ws.CLOSING);
  ws.onclose = function(e) {
    assert_equals(ws.readyState, ws.CLOSED);
    assert_equals(e.wasClean, false);
    done();
  };
}, 1000);
ws.onopen = ws.onclose = assert_unreached;
};
</script>
