Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/websockets/interfaces/WebSocket/close/close-connecting-expected.txt


FAIL WebSockets: close() when connecting assert_equals: expected 2 but got 3
